+++ config.js
@@ -23,7 +23,7 @@ var conf = {
 };
 
 conf.theme = {
-    name: 'basic'
+    name: 'wood'
 };
 
 /* third party settings */